refine migration script
git-svn-id: https://svn.nixos.org/repos/nix/services/trunk@2925 70bd8c7a-acb8-0310-9f0d-9cc1c95dcdbb

+++ twiki/migrate-data.sh
@@ -12,15 +12,28 @@ do
   do 
      cp -r --reply=yes $file $target
   done
-  mv $target/WebContents.txt   $target/WebLeftBar.txt
-  mv $target/WebContents.txt,v $target/WebLeftBar.txt,v
+  cp $source/.changes $source/.mailnotify $target
+  if test -f $target/WebContents.txt
+  then
+    mv $target/WebContents.txt   $target/WebLeftBar.txt
+    mv $target/WebContents.txt,v $target/WebLeftBar.txt,v
+  fi
 done
 
+echo "Copying password file ..."
 cp -r --reply=yes $old/data/TWiki/CustomSiteMenus* $new/data/TWiki/
 
-cp -r --reply=yes $old/pub/* $new/pub/
-
+echo "Copying password file ..."
 cp --reply=yes $old/data/.htpasswd $new/data/
+
+echo "Copying log files ..."
+cp --reply=yes $old/data/log* $new/data/
+
+echo "Touching default web topics ..."
+for dir in `find $new/data -type d`
+do
+  touch $dir/WebCustomMenus.txt $dir/WebNews.txt
+done
 
 echo "Patching RCS locks ..."
 for file in `find $new/data -name \*.txt,v`
@@ -30,8 +43,5 @@ do
   mv $file.copy $file
 done
 
-echo "Touching default web topics ..."
-for dir in `find $new/data -type d`
-do
-  touch $dir/WebCustomMenus.txt $dir/WebNews.txt
-done
+echo "Copying pub/ ..."
+cp -r --reply=yes $old/pub/* $new/pub/
